+++ rundeckapp/grails-spa/src/utilities/RundeckVersion.ts
